% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_weibullPH.R
\name{sim_weibullPH}
\alias{sim_weibullPH}
\title{Function to simulate Weibull with proportional hazard parametrisation}
\usage{
sim_weibullPH(
  lp,
  shape,
  scale,
  t_max,
  t_entry = rep(0, length(lp)),
  lp_cure = NULL,
  lp_cen = NULL
)
}
\arguments{
\item{lp}{`<numeric()>` a vector a linear predictor i.e. sum of log hazard-ratio}

\item{shape}{`<integer(1)>`Weibull shape parameter, positive}

\item{scale}{`<integer(1)>` Weibull scale parameter,  positive}

\item{t_max}{`<numeric(1)>` maximium follow-up time}

\item{t_entry}{`<numeric()>`vector of entry time}

\item{lp_cure}{`<numeric()>` vector of log-odds of being cured/immortal. If a scalar, everyone shares the same immortal rate}

\item{lp_cen}{`<numeric()>` vector of log-hazard of dropping out. If a scalar, everyone shares the same immortal rate}
}
\value{
a `data.frame` with these columns: tt_ev, ev, lp
}
\description{
Function to simulate Weibull with proportional hazard parametrisation
}
