% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sim_weibullPH.R
\name{surv}
\alias{surv}
\title{Baseline hazard
Real hazard at t
Survival at t = exp(-H) = exp(int_t(h(t)))}
\usage{
surv(t, lp, shape, scale)
}
\description{
Baseline hazard
Real hazard at t
Survival at t = exp(-H) = exp(int_t(h(t)))
}
